+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.9] - 2019-12-29
+### Added
+- Logging initial configuration (later, the interface will be injected in application services);
+- Explicit URL binding on initialization (Docker port binding);
+- Worker project backbone, responsible for queue consumption at a later development stage;
+- Some documentation.
+
+### Fixed
+- Docker file runtime image.
+
 ## [0.0.8] - 2019-12-29
 ### Added
 - API's Dockerfile.